added tags menu and merged menu controllers
